refactor(rule-view): move and rename rule-view component files
