+++ frontend/src/app/rule-management/app-rule/components/rule-view/rule-view.component.html
